requirements.txt
numpy-stl
Pillow
trimesh
progressbar2